api/requirements.txt
azure-functions
pyodbc
azure-identity
